refactor(gui): put a transport under the editor bridge
`GuiBridge` reached the editor through an `mpsc::UnboundedSender<GuiRequest>`
it owned directly, so the only editor it could ever address was one on a
thread in this process.

Introduce `GuiTransport`: send a `GuiCommand` and await the `GuiReply` its
`reply_kind` promises, plus subscribe to a watch stream of snapshots. The
existing channel machinery becomes `LocalTransport`, and `GuiBridge` is now a
thin `Arc<dyn GuiTransport>` with the same 31 typed helpers, so the Tauri
command layer is untouched.

The future is boxed rather than written as `async fn` in the trait: an
`async fn` in a trait is not dyn-compatible, and `ovim-core` already spells
its object-safe async traits this way, so no new dependency is needed.

`shutdown` runs from Tauri's synchronous exit handler, so the trait carries a
separate one-way send for commands that answer nothing, and rejects anything
whose answer would then be dropped in silence.

Moving the bridge into its own module also brings `gui/mod.rs` back under the
3k-line refactor threshold: 3108 non-test lines to 2258.

Tests drive a bridge over a scripted local editor for every reply shape
including fire-and-forget shutdown, replay every typed helper against a
recording transport to catch a helper wired to the wrong variant, and pin the
"editor stopped" and "closed the response" wordings that reach the user.
