+++ ovim/src/gui/mod.rs
@@ -5,9 +5,15 @@
 //! intentionally not an HTML text editor pretending to be Vim: keyboard,
 //! paste, picker, tab, and file-tree actions all flow back through Ovim's real
 //! input/state machinery.
+//!
+//! This module holds the editor loop and the projection from editor state to
+//! [`GuiSnapshot`]. The conversation with that editor -- `GuiBridge` and the
+//! transport it runs over -- lives in [`bridge`], and the wire types both ends
+//! share live in [`protocol`].
 
 #[cfg(feature = "gui")]
 pub mod app;
+pub mod bridge;
 #[cfg(feature = "gui")]
 pub mod browser;
 #[cfg(feature = "gui")]
@@ -16,9 +22,14 @@ pub mod protocol;
 #[cfg(feature = "gui")]
 pub mod window;
 
-// The protocol types used to live in this module. They are re-exported so the
-// Tauri command layer and the API keep importing them from `gui`, and so that
-// `gui::protocol` stays the single place a transport has to look.
+// The protocol types and the bridge used to live in this module. They are
+// re-exported so the Tauri command layer and the API keep importing them from
+// `gui`, and so that `gui::protocol` stays the single place a transport has to
+// look.
+pub use bridge::{
+    GuiBridge, GuiReplyReceiver, GuiReplySender, GuiRequest, GuiTransport, GuiTransportFuture,
+    LocalTransport,
+};
 pub use protocol::{
     GuiAgentOption, GuiAiChat, GuiAiProfileOption, GuiChatMessage, GuiChatSetup,
     GuiChatSetupAction, GuiCodeAttachment, GuiCodeExplanation, GuiCodeExplanationDiscussion,
@@ -45,7 +56,7 @@ use anyhow::{Context, Result};
 use std::path::Path;
 use std::sync::mpsc as std_mpsc;
 use std::time::{Duration, Instant};
-use tokio::sync::{mpsc, oneshot, watch};
+use tokio::sync::{mpsc, watch};
 use unicode_segmentation::UnicodeSegmentation;
 
 const TICK_RATE: Duration = Duration::from_millis(50);
@@ -189,867 +200,6 @@ fn attach_chat_images(editor: &mut Editor, paths: &[std::path::PathBuf]) -> Resu
         editor.handle_paste_event(path)?;
     }
     Ok(())
-}
-
-/// One editor request, with the channel its answer travels back on.
-///
-/// This is the in-process form: it pairs the serializable
-/// [`GuiCommand`] payload with a `oneshot::Sender`, and
-/// [`GuiRequest::into_parts`] separates the two again for a transport that
-/// cannot carry a channel.
-pub enum GuiRequest {
-    Snapshot {
-        columns: u16,
-        rows: u16,
-        reply: oneshot::Sender<Result<GuiSnapshot, String>>,
-    },
-    VectorSource {
-        reply: oneshot::Sender<Result<GuiVectorSource, String>>,
-    },
-    VectorFeedback {
-        feedback: String,
-        reply: oneshot::Sender<Result<(), String>>,
-    },
-    DiffWorkspace {
-        reply: oneshot::Sender<Result<std::path::PathBuf, String>>,
-    },
-    OpenDiffBuffer {
-        title: String,
-        content: String,
-        reply: oneshot::Sender<Result<(), String>>,
-    },
-    Key {
-        input: GuiKeyInput,
-        reply: oneshot::Sender<Result<(), String>>,
-    },
-    OpenAiChat {
-        reply: oneshot::Sender<Result<(), String>>,
-    },
-    UpdateChatInput {
-        expected_input: String,
-        expected_cursor: usize,
-        input: String,
-        cursor: usize,
-        action: Option<GuiKeyInput>,
-        reply: oneshot::Sender<Result<(), String>>,
-    },
-    SetChatInputCursor {
-        offset: usize,
-        reply: oneshot::Sender<Result<(), String>>,
-    },
-    SetChatInputWidth {
-        columns: usize,
-        reply: oneshot::Sender<Result<(), String>>,
-    },
-    RemoveChatImage {
-        index: usize,
-        reply: oneshot::Sender<Result<(), String>>,
-    },
-    SelectAiProfile {
-        profile: String,
-        reply: oneshot::Sender<Result<(), String>>,
-    },
-    SelectReasoningEffort {
-        effort: String,
-        reply: oneshot::Sender<Result<(), String>>,
-    },
-    SelectChatMessage {
-        index: usize,
-        reply: oneshot::Sender<Result<(), String>>,
-    },
-    ManageQueuedChatInput {
-        id: u64,
-        action: String,
-        reply: oneshot::Sender<Result<(), String>>,
-    },
-    AiPolicy {
-        action: String,
-        reply: oneshot::Sender<Result<(), String>>,
-    },
-    EditorCommand {
-        command: String,
-        reply: oneshot::Sender<Result<(), String>>,
-    },
-    SelectChatAgent {
-        agent_id: Option<String>,
-        reply: oneshot::Sender<Result<(), String>>,
-    },
-    Paste {
-        text: String,
-        reply: oneshot::Sender<Result<(), String>>,
-    },
-    AttachImages {
-        paths: Vec<std::path::PathBuf>,
-        reply: oneshot::Sender<Result<(), String>>,
-    },
-    AttachImageData {
-        name: String,
-        data: Vec<u8>,
-        reply: oneshot::Sender<Result<(), String>>,
-    },
-    SetCursor {
-        pane: usize,
-        line: usize,
-        display_column: usize,
-        reply: oneshot::Sender<Result<(), String>>,
-    },
-    SelectTab {
-        index: usize,
-        reply: oneshot::Sender<Result<(), String>>,
-    },
-    FocusPane {
-        index: usize,
-        reply: oneshot::Sender<Result<(), String>>,
-    },
-    SelectPicker {
-        index: usize,
-        reply: oneshot::Sender<Result<(), String>>,
-    },
-    SelectCompletion {
-        index: usize,
-        activate: bool,
-        reply: oneshot::Sender<Result<(), String>>,
-    },
-    SelectFileTree {
-        index: usize,
-        activate: bool,
-        reply: oneshot::Sender<Result<(), String>>,
-    },
-    SelectProblem {
-        kind: String,
-        index: usize,
-        activate: bool,
-        reply: oneshot::Sender<Result<(), String>>,
-    },
-    SelectLsp {
-        index: usize,
-        activate: bool,
-        reply: oneshot::Sender<Result<(), String>>,
-    },
-    SelectDebugFrame {
-        index: usize,
-        reply: oneshot::Sender<Result<(), String>>,
-    },
-    Shutdown,
-}
-
-/// The reply half of a [`GuiRequest`], separated from its payload.
-///
-/// A [`GuiCommand`] can cross a process boundary but a `oneshot::Sender`
-/// cannot, so the two travel apart and are rejoined by
-/// [`GuiRequest::from_parts`] on the editor's side of the transport.
-pub enum GuiReplySender {
-    None,
-    Unit(oneshot::Sender<Result<(), String>>),
-    Snapshot(oneshot::Sender<Result<GuiSnapshot, String>>),
-    VectorSource(oneshot::Sender<Result<GuiVectorSource, String>>),
-    Path(oneshot::Sender<Result<std::path::PathBuf, String>>),
-}
-
-impl GuiReplySender {
-    /// Build the reply channel that `kind` describes.
-    ///
-    /// Callers get the shape from [`GuiCommand::reply_kind`] rather than
-    /// choosing it, which is what makes [`GuiRequest::from_parts`] unable to
-    /// pair a command with the wrong channel in practice.
-    pub fn channel(kind: GuiReplyKind) -> (Self, GuiReplyReceiver) {
-        match kind {
-            GuiReplyKind::None => (GuiReplySender::None, GuiReplyReceiver::None),
-            GuiReplyKind::Unit => {
-                let (tx, rx) = oneshot::channel();
-                (GuiReplySender::Unit(tx), GuiReplyReceiver::Unit(rx))
-            }
-            GuiReplyKind::Snapshot => {
-                let (tx, rx) = oneshot::channel();
-                (GuiReplySender::Snapshot(tx), GuiReplyReceiver::Snapshot(rx))
-            }
-            GuiReplyKind::VectorSource => {
-                let (tx, rx) = oneshot::channel();
-                (
-                    GuiReplySender::VectorSource(tx),
-                    GuiReplyReceiver::VectorSource(rx),
-                )
-            }
-            GuiReplyKind::Path => {
-                let (tx, rx) = oneshot::channel();
-                (GuiReplySender::Path(tx), GuiReplyReceiver::Path(rx))
-            }
-        }
-    }
-
-    pub fn kind(&self) -> GuiReplyKind {
-        match self {
-            GuiReplySender::None => GuiReplyKind::None,
-            GuiReplySender::Unit(_) => GuiReplyKind::Unit,
-            GuiReplySender::Snapshot(_) => GuiReplyKind::Snapshot,
-            GuiReplySender::VectorSource(_) => GuiReplyKind::VectorSource,
-            GuiReplySender::Path(_) => GuiReplyKind::Path,
-        }
-    }
-}
-
-/// The receiving end of a [`GuiReplySender`].
-pub enum GuiReplyReceiver {
-    None,
-    Unit(oneshot::Receiver<Result<(), String>>),
-    Snapshot(oneshot::Receiver<Result<GuiSnapshot, String>>),
-    VectorSource(oneshot::Receiver<Result<GuiVectorSource, String>>),
-    Path(oneshot::Receiver<Result<std::path::PathBuf, String>>),
-}
-
-impl GuiReplyReceiver {
-    /// Await the editor's answer.
-    ///
-    /// `Ok(None)` means the command is fire-and-forget and there is nothing to
-    /// wait for; `Err` means the editor dropped the channel without answering.
-    pub async fn recv(self) -> Result<Option<GuiReply>, String> {
-        const CLOSED: &str = "The Ovim editor thread closed the response";
-        Ok(match self {
-            GuiReplyReceiver::None => None,
-            GuiReplyReceiver::Unit(rx) => {
-                Some(GuiReply::Unit(rx.await.map_err(|_| CLOSED.to_string())?))
-            }
-            GuiReplyReceiver::Snapshot(rx) => Some(GuiReply::Snapshot(Box::new(
-                rx.await.map_err(|_| CLOSED.to_string())?,
-            ))),
-            GuiReplyReceiver::VectorSource(rx) => Some(GuiReply::VectorSource(
-                rx.await.map_err(|_| CLOSED.to_string())?,
-            )),
-            GuiReplyReceiver::Path(rx) => {
-                Some(GuiReply::Path(rx.await.map_err(|_| CLOSED.to_string())?))
-            }
-        })
-    }
-}
-
-impl GuiRequest {
-    /// Rejoin a wire payload with a reply channel.
-    ///
-    /// The channel has to match [`GuiCommand::reply_kind`]. A caller that
-    /// builds it with [`GuiReplySender::channel`] cannot get that wrong, so a
-    /// mismatch is a transport bug and is reported rather than papered over by
-    /// answering on a channel nobody is waiting on.
-    pub fn from_parts(command: GuiCommand, reply: GuiReplySender) -> Result<Self, String> {
-        Ok(match (command, reply) {
-            (GuiCommand::Snapshot { columns, rows }, GuiReplySender::Snapshot(reply)) => {
-                GuiRequest::Snapshot {
-                    columns,
-                    rows,
-                    reply,
-                }
-            }
-            (GuiCommand::VectorSource, GuiReplySender::VectorSource(reply)) => {
-                GuiRequest::VectorSource { reply }
-            }
-            (GuiCommand::VectorFeedback { feedback }, GuiReplySender::Unit(reply)) => {
-                GuiRequest::VectorFeedback { feedback, reply }
-            }
-            (GuiCommand::DiffWorkspace, GuiReplySender::Path(reply)) => {
-                GuiRequest::DiffWorkspace { reply }
-            }
-            (GuiCommand::OpenDiffBuffer { title, content }, GuiReplySender::Unit(reply)) => {
-                GuiRequest::OpenDiffBuffer {
-                    title,
-                    content,
-                    reply,
-                }
-            }
-            (GuiCommand::Key { input }, GuiReplySender::Unit(reply)) => {
-                GuiRequest::Key { input, reply }
-            }
-            (GuiCommand::OpenAiChat, GuiReplySender::Unit(reply)) => {
-                GuiRequest::OpenAiChat { reply }
-            }
-            (
-                GuiCommand::UpdateChatInput {
-                    expected_input,
-                    expected_cursor,
-                    input,
-                    cursor,
-                    action,
-                },
-                GuiReplySender::Unit(reply),
-            ) => GuiRequest::UpdateChatInput {
-                expected_input,
-                expected_cursor,
-                input,
-                cursor,
-                action,
-                reply,
-            },
-            (GuiCommand::SetChatInputCursor { offset }, GuiReplySender::Unit(reply)) => {
-                GuiRequest::SetChatInputCursor { offset, reply }
-            }
-            (GuiCommand::SetChatInputWidth { columns }, GuiReplySender::Unit(reply)) => {
-                GuiRequest::SetChatInputWidth { columns, reply }
-            }
-            (GuiCommand::RemoveChatImage { index }, GuiReplySender::Unit(reply)) => {
-                GuiRequest::RemoveChatImage { index, reply }
-            }
-            (GuiCommand::SelectAiProfile { profile }, GuiReplySender::Unit(reply)) => {
-                GuiRequest::SelectAiProfile { profile, reply }
-            }
-            (GuiCommand::SelectReasoningEffort { effort }, GuiReplySender::Unit(reply)) => {
-                GuiRequest::SelectReasoningEffort { effort, reply }
-            }
-            (GuiCommand::SelectChatMessage { index }, GuiReplySender::Unit(reply)) => {
-                GuiRequest::SelectChatMessage { index, reply }
-            }
-            (GuiCommand::ManageQueuedChatInput { id, action }, GuiReplySender::Unit(reply)) => {
-                GuiRequest::ManageQueuedChatInput { id, action, reply }
-            }
-            (GuiCommand::AiPolicy { action }, GuiReplySender::Unit(reply)) => {
-                GuiRequest::AiPolicy { action, reply }
-            }
-            (GuiCommand::EditorCommand { command }, GuiReplySender::Unit(reply)) => {
-                GuiRequest::EditorCommand { command, reply }
-            }
-            (GuiCommand::SelectChatAgent { agent_id }, GuiReplySender::Unit(reply)) => {
-                GuiRequest::SelectChatAgent { agent_id, reply }
-            }
-            (GuiCommand::Paste { text }, GuiReplySender::Unit(reply)) => {
-                GuiRequest::Paste { text, reply }
-            }
-            (GuiCommand::AttachImages { paths }, GuiReplySender::Unit(reply)) => {
-                GuiRequest::AttachImages { paths, reply }
-            }
-            (GuiCommand::AttachImageData { name, data }, GuiReplySender::Unit(reply)) => {
-                GuiRequest::AttachImageData { name, data, reply }
-            }
-            (
-                GuiCommand::SetCursor {
-                    pane,
-                    line,
-                    display_column,
-                },
-                GuiReplySender::Unit(reply),
-            ) => GuiRequest::SetCursor {
-                pane,
-                line,
-                display_column,
-                reply,
-            },
-            (GuiCommand::SelectTab { index }, GuiReplySender::Unit(reply)) => {
-                GuiRequest::SelectTab { index, reply }
-            }
-            (GuiCommand::FocusPane { index }, GuiReplySender::Unit(reply)) => {
-                GuiRequest::FocusPane { index, reply }
-            }
-            (GuiCommand::SelectPicker { index }, GuiReplySender::Unit(reply)) => {
-                GuiRequest::SelectPicker { index, reply }
-            }
-            (GuiCommand::SelectCompletion { index, activate }, GuiReplySender::Unit(reply)) => {
-                GuiRequest::SelectCompletion {
-                    index,
-                    activate,
-                    reply,
-                }
-            }
-            (GuiCommand::SelectFileTree { index, activate }, GuiReplySender::Unit(reply)) => {
-                GuiRequest::SelectFileTree {
-                    index,
-                    activate,
-                    reply,
-                }
-            }
-            (
-                GuiCommand::SelectProblem {
-                    kind,
-                    index,
-                    activate,
-                },
-                GuiReplySender::Unit(reply),
-            ) => GuiRequest::SelectProblem {
-                kind,
-                index,
-                activate,
-                reply,
-            },
-            (GuiCommand::SelectLsp { index, activate }, GuiReplySender::Unit(reply)) => {
-                GuiRequest::SelectLsp {
-                    index,
-                    activate,
-                    reply,
-                }
-            }
-            (GuiCommand::SelectDebugFrame { index }, GuiReplySender::Unit(reply)) => {
-                GuiRequest::SelectDebugFrame { index, reply }
-            }
-            (GuiCommand::Shutdown, GuiReplySender::None) => GuiRequest::Shutdown,
-            (command, reply) => {
-                return Err(format!(
-                    "A GUI command needing a {:?} reply was given a {:?} reply channel",
-                    command.reply_kind(),
-                    reply.kind()
-                ))
-            }
-        })
-    }
-
-    /// Split a request back into the payload a transport can serialize and the
-    /// reply channel it cannot.
-    ///
-    /// This is the exact inverse of [`GuiRequest::from_parts`]; the round-trip
-    /// test in this module is what keeps a field from being dropped on the way
-    /// through the wire format.
-    pub fn into_parts(self) -> (GuiCommand, GuiReplySender) {
-        match self {
-            GuiRequest::Snapshot {
-                columns,
-                rows,
-                reply,
-            } => (
-                GuiCommand::Snapshot { columns, rows },
-                GuiReplySender::Snapshot(reply),
-            ),
-            GuiRequest::VectorSource { reply } => (
-                GuiCommand::VectorSource,
-                GuiReplySender::VectorSource(reply),
-            ),
-            GuiRequest::VectorFeedback { feedback, reply } => (
-                GuiCommand::VectorFeedback { feedback },
-                GuiReplySender::Unit(reply),
-            ),
-            GuiRequest::DiffWorkspace { reply } => {
-                (GuiCommand::DiffWorkspace, GuiReplySender::Path(reply))
-            }
-            GuiRequest::OpenDiffBuffer {
-                title,
-                content,
-                reply,
-            } => (
-                GuiCommand::OpenDiffBuffer { title, content },
-                GuiReplySender::Unit(reply),
-            ),
-            GuiRequest::Key { input, reply } => {
-                (GuiCommand::Key { input }, GuiReplySender::Unit(reply))
-            }
-            GuiRequest::OpenAiChat { reply } => {
-                (GuiCommand::OpenAiChat, GuiReplySender::Unit(reply))
-            }
-            GuiRequest::UpdateChatInput {
-                expected_input,
-                expected_cursor,
-                input,
-                cursor,
-                action,
-                reply,
-            } => (
-                GuiCommand::UpdateChatInput {
-                    expected_input,
-                    expected_cursor,
-                    input,
-                    cursor,
-                    action,
-                },
-                GuiReplySender::Unit(reply),
-            ),
-            GuiRequest::SetChatInputCursor { offset, reply } => (
-                GuiCommand::SetChatInputCursor { offset },
-                GuiReplySender::Unit(reply),
-            ),
-            GuiRequest::SetChatInputWidth { columns, reply } => (
-                GuiCommand::SetChatInputWidth { columns },
-                GuiReplySender::Unit(reply),
-            ),
-            GuiRequest::RemoveChatImage { index, reply } => (
-                GuiCommand::RemoveChatImage { index },
-                GuiReplySender::Unit(reply),
-            ),
-            GuiRequest::SelectAiProfile { profile, reply } => (
-                GuiCommand::SelectAiProfile { profile },
-                GuiReplySender::Unit(reply),
-            ),
-            GuiRequest::SelectReasoningEffort { effort, reply } => (
-                GuiCommand::SelectReasoningEffort { effort },
-                GuiReplySender::Unit(reply),
-            ),
-            GuiRequest::SelectChatMessage { index, reply } => (
-                GuiCommand::SelectChatMessage { index },
-                GuiReplySender::Unit(reply),
-            ),
-            GuiRequest::ManageQueuedChatInput { id, action, reply } => (
-                GuiCommand::ManageQueuedChatInput { id, action },
-                GuiReplySender::Unit(reply),
-            ),
-            GuiRequest::AiPolicy { action, reply } => {
-                (GuiCommand::AiPolicy { action }, GuiReplySender::Unit(reply))
-            }
-            GuiRequest::EditorCommand { command, reply } => (
-                GuiCommand::EditorCommand { command },
-                GuiReplySender::Unit(reply),
-            ),
-            GuiRequest::SelectChatAgent { agent_id, reply } => (
-                GuiCommand::SelectChatAgent { agent_id },
-                GuiReplySender::Unit(reply),
-            ),
-            GuiRequest::Paste { text, reply } => {
-                (GuiCommand::Paste { text }, GuiReplySender::Unit(reply))
-            }
-            GuiRequest::AttachImages { paths, reply } => (
-                GuiCommand::AttachImages { paths },
-                GuiReplySender::Unit(reply),
-            ),
-            GuiRequest::AttachImageData { name, data, reply } => (
-                GuiCommand::AttachImageData { name, data },
-                GuiReplySender::Unit(reply),
-            ),
-            GuiRequest::SetCursor {
-                pane,
-                line,
-                display_column,
-                reply,
-            } => (
-                GuiCommand::SetCursor {
-                    pane,
-                    line,
-                    display_column,
-                },
-                GuiReplySender::Unit(reply),
-            ),
-            GuiRequest::SelectTab { index, reply } => {
-                (GuiCommand::SelectTab { index }, GuiReplySender::Unit(reply))
-            }
-            GuiRequest::FocusPane { index, reply } => {
-                (GuiCommand::FocusPane { index }, GuiReplySender::Unit(reply))
-            }
-            GuiRequest::SelectPicker { index, reply } => (
-                GuiCommand::SelectPicker { index },
-                GuiReplySender::Unit(reply),
-            ),
-            GuiRequest::SelectCompletion {
-                index,
-                activate,
-                reply,
-            } => (
-                GuiCommand::SelectCompletion { index, activate },
-                GuiReplySender::Unit(reply),
-            ),
-            GuiRequest::SelectFileTree {
-                index,
-                activate,
-                reply,
-            } => (
-                GuiCommand::SelectFileTree { index, activate },
-                GuiReplySender::Unit(reply),
-            ),
-            GuiRequest::SelectProblem {
-                kind,
-                index,
-                activate,
-                reply,
-            } => (
-                GuiCommand::SelectProblem {
-                    kind,
-                    index,
-                    activate,
-                },
-                GuiReplySender::Unit(reply),
-            ),
-            GuiRequest::SelectLsp {
-                index,
-                activate,
-                reply,
-            } => (
-                GuiCommand::SelectLsp { index, activate },
-                GuiReplySender::Unit(reply),
-            ),
-            GuiRequest::SelectDebugFrame { index, reply } => (
-                GuiCommand::SelectDebugFrame { index },
-                GuiReplySender::Unit(reply),
-            ),
-            GuiRequest::Shutdown => (GuiCommand::Shutdown, GuiReplySender::None),
-        }
-    }
-}
-
-/// Send-side handle stored as Tauri application state.
-#[derive(Clone)]
-pub struct GuiBridge {
-    requests: mpsc::UnboundedSender<GuiRequest>,
-    updates: watch::Sender<Option<GuiSnapshot>>,
-}
-
-impl GuiBridge {
-    pub fn spawn(file: Option<FileArg>, resume: bool, services: EditorServices) -> Result<Self> {
-        let (request_tx, request_rx) = mpsc::unbounded_channel();
-        let (update_tx, _) = watch::channel(None);
-        let (ready_tx, ready_rx) = std_mpsc::sync_channel(1);
-        let editor_updates = update_tx.clone();
-
-        std::thread::Builder::new()
-            .name("ovim-gui-editor".to_string())
-            .spawn(move || {
-                let runtime = tokio::runtime::Builder::new_multi_thread()
-                    .worker_threads(2)
-                    .enable_all()
-                    .build();
-                let Ok(runtime) = runtime else {
-                    let _ = ready_tx.send(Err("Failed to create GUI runtime".to_string()));
-                    return;
-                };
-                runtime.block_on(run_editor(
-                    file,
-                    resume,
-                    services,
-                    request_rx,
-                    editor_updates,
-                    ready_tx,
-                ));
-            })
-            .context("Failed to start the GUI editor thread")?;
-
-        match ready_rx.recv_timeout(Duration::from_secs(15)) {
-            Ok(Ok(())) => Ok(Self {
-                requests: request_tx,
-                updates: update_tx,
-            }),
-            Ok(Err(error)) => anyhow::bail!(error),
-            Err(error) => anyhow::bail!("GUI editor initialization timed out: {error}"),
-        }
-    }
-
-    /// Subscribe to coalesced editor-state changes.
-    ///
-    /// Tauri turns this watch stream into an IPC channel. Slow webviews only
-    /// retain the newest snapshot instead of building an unbounded queue.
-    pub fn subscribe(&self) -> watch::Receiver<Option<GuiSnapshot>> {
-        self.updates.subscribe()
-    }
-
-    pub async fn snapshot(&self, columns: u16, rows: u16) -> Result<GuiSnapshot, String> {
-        let (reply, response) = oneshot::channel();
-        self.requests
-            .send(GuiRequest::Snapshot {
-                columns,
-                rows,
-                reply,
-            })
-            .map_err(|_| "The Ovim editor thread has stopped".to_string())?;
-        response
-            .await
-            .map_err(|_| "The Ovim editor thread closed the response".to_string())?
-    }
-
-    pub async fn vector_source(&self) -> Result<GuiVectorSource, String> {
-        let (reply, response) = oneshot::channel();
-        self.requests
-            .send(GuiRequest::VectorSource { reply })
-            .map_err(|_| "The Ovim editor thread has stopped".to_string())?;
-        response
-            .await
-            .map_err(|_| "The Ovim editor thread closed the response".to_string())?
-    }
-
-    pub async fn vector_feedback(&self, feedback: String) -> Result<(), String> {
-        self.request(|reply| GuiRequest::VectorFeedback { feedback, reply })
-            .await
-    }
-
-    pub async fn diff_workspace(&self) -> Result<std::path::PathBuf, String> {
-        let (reply, response) = oneshot::channel();
-        self.requests
-            .send(GuiRequest::DiffWorkspace { reply })
-            .map_err(|_| "The Ovim editor thread has stopped".to_string())?;
-        response
-            .await
-            .map_err(|_| "The Ovim editor thread closed the response".to_string())?
-    }
-
-    pub async fn open_diff_buffer(&self, title: String, content: String) -> Result<(), String> {
-        self.request(|reply| GuiRequest::OpenDiffBuffer {
-            title,
-            content,
-            reply,
-        })
-        .await
-    }
-
-    pub async fn key(&self, input: GuiKeyInput) -> Result<(), String> {
-        self.request(|reply| GuiRequest::Key { input, reply }).await
-    }
-
-    pub async fn open_ai_chat(&self) -> Result<(), String> {
-        self.request(|reply| GuiRequest::OpenAiChat { reply }).await
-    }
-
-    pub async fn paste(&self, text: String) -> Result<(), String> {
-        self.request(|reply| GuiRequest::Paste { text, reply })
-            .await
-    }
-
-    pub async fn set_chat_input_cursor(&self, offset: usize) -> Result<(), String> {
-        self.request(|reply| GuiRequest::SetChatInputCursor { offset, reply })
-            .await
-    }
-
-    pub async fn update_chat_input(
-        &self,
-        expected_input: String,
-        expected_cursor: usize,
-        input: String,
-        cursor: usize,
-        action: Option<GuiKeyInput>,
-    ) -> Result<(), String> {
-        self.request(|reply| GuiRequest::UpdateChatInput {
-            expected_input,
-            expected_cursor,
-            input,
-            cursor,
-            action,
-            reply,
-        })
-        .await
-    }
-
-    pub async fn set_chat_input_width(&self, columns: usize) -> Result<(), String> {
-        self.request(|reply| GuiRequest::SetChatInputWidth { columns, reply })
-            .await
-    }
-
-    pub async fn remove_chat_image(&self, index: usize) -> Result<(), String> {
-        self.request(|reply| GuiRequest::RemoveChatImage { index, reply })
-            .await
-    }
-
-    pub async fn attach_images(&self, paths: Vec<std::path::PathBuf>) -> Result<(), String> {
-        self.request(|reply| GuiRequest::AttachImages { paths, reply })
-            .await
-    }
-
-    pub async fn attach_image_data(&self, name: String, data: Vec<u8>) -> Result<(), String> {
-        self.request(|reply| GuiRequest::AttachImageData { name, data, reply })
-            .await
-    }
-
-    pub async fn set_cursor(
-        &self,
-        pane: usize,
-        line: usize,
-        display_column: usize,
-    ) -> Result<(), String> {
-        self.request(|reply| GuiRequest::SetCursor {
-            pane,
-            line,
-            display_column,
-            reply,
-        })
-        .await
-    }
-
-    pub async fn select_tab(&self, index: usize) -> Result<(), String> {
-        self.request(|reply| GuiRequest::SelectTab { index, reply })
-            .await
-    }
-
-    pub async fn focus_pane(&self, index: usize) -> Result<(), String> {
-        self.request(|reply| GuiRequest::FocusPane { index, reply })
-            .await
-    }
-
-    pub async fn select_picker(&self, index: usize) -> Result<(), String> {
-        self.request(|reply| GuiRequest::SelectPicker { index, reply })
-            .await
-    }
-
-    pub async fn select_completion(&self, index: usize, activate: bool) -> Result<(), String> {
-        self.request(|reply| GuiRequest::SelectCompletion {
-            index,
-            activate,
-            reply,
-        })
-        .await
-    }
-
-    pub async fn select_file_tree(&self, index: usize, activate: bool) -> Result<(), String> {
-        self.request(|reply| GuiRequest::SelectFileTree {
-            index,
-            activate,
-            reply,
-        })
-        .await
-    }
-
-    pub async fn select_problem(
-        &self,
-        kind: String,
-        index: usize,
-        activate: bool,
-    ) -> Result<(), String> {
-        self.request(|reply| GuiRequest::SelectProblem {
-            kind,
-            index,
-            activate,
-            reply,
-        })
-        .await
-    }
-
-    pub async fn select_lsp(&self, index: usize, activate: bool) -> Result<(), String> {
-        self.request(|reply| GuiRequest::SelectLsp {
-            index,
-            activate,
-            reply,
-        })
-        .await
-    }
-
-    pub async fn select_debug_frame(&self, index: usize) -> Result<(), String> {
-        self.request(|reply| GuiRequest::SelectDebugFrame { index, reply })
-            .await
-    }
-
-    pub async fn select_ai_profile(&self, profile: String) -> Result<(), String> {
-        self.request(|reply| GuiRequest::SelectAiProfile { profile, reply })
-            .await
-    }
-
-    pub async fn select_reasoning_effort(&self, effort: String) -> Result<(), String> {
-        self.request(|reply| GuiRequest::SelectReasoningEffort { effort, reply })
-            .await
-    }
-
-    pub async fn select_chat_message(&self, index: usize) -> Result<(), String> {
-        self.request(|reply| GuiRequest::SelectChatMessage { index, reply })
-            .await
-    }
-
-    pub async fn manage_queued_chat_input(&self, id: u64, action: String) -> Result<(), String> {
-        self.request(|reply| GuiRequest::ManageQueuedChatInput { id, action, reply })
-            .await
-    }
-
-    pub async fn ai_policy(&self, action: String) -> Result<(), String> {
-        self.request(|reply| GuiRequest::AiPolicy { action, reply })
-            .await
-    }
-
-    pub async fn editor_command(&self, command: String) -> Result<(), String> {
-        self.request(|reply| GuiRequest::EditorCommand { command, reply })
-            .await
-    }
-
-    pub async fn select_chat_agent(&self, agent_id: Option<String>) -> Result<(), String> {
-        self.request(|reply| GuiRequest::SelectChatAgent { agent_id, reply })
-            .await
-    }
-
-    pub fn shutdown(&self) {
-        let _ = self.requests.send(GuiRequest::Shutdown);
-    }
-
-    async fn request(
-        &self,
-        request: impl FnOnce(oneshot::Sender<Result<(), String>>) -> GuiRequest,
-    ) -> Result<(), String> {
-        let (reply_tx, reply_rx) = oneshot::channel();
-        self.requests
-            .send(request(reply_tx))
-            .map_err(|_| "The Ovim editor thread has stopped".to_string())?;
-        reply_rx
-            .await
-            .map_err(|_| "The Ovim editor thread closed the response".to_string())?
-    }
 }
 
 async fn run_editor(
@@ -3167,72 +2317,6 @@ mod tests {
         assert!(json["gitChanges"]["added"].is_number());
         assert_eq!(json["layout"]["kind"], "pane");
         assert!(json.get("file_name").is_none());
-    }
-
-    #[test]
-    fn every_command_rebuilds_the_request_it_came_from_without_losing_a_field() {
-        // `into_parts` is the inverse of `from_parts`, so a payload field that
-        // either conversion forgets shows up as an inequality here.
-        for command in protocol::sample_commands() {
-            let (reply, _receiver) = GuiReplySender::channel(command.reply_kind());
-            let request = GuiRequest::from_parts(command.clone(), reply)
-                .unwrap_or_else(|error| panic!("{command:?} could not be rebuilt: {error}"));
-
-            let (recovered, reply) = request.into_parts();
-
-            assert_eq!(recovered, command);
-            assert_eq!(reply.kind(), command.reply_kind());
-        }
-    }
-
-    #[test]
-    fn a_reply_channel_of_the_wrong_shape_is_refused_rather_than_mis_wired() {
-        let (reply, _receiver) = GuiReplySender::channel(GuiReplyKind::Unit);
-
-        let error = GuiRequest::from_parts(
-            GuiCommand::Snapshot {
-                columns: 80,
-                rows: 24,
-            },
-            reply,
-        )
-        .err()
-        .expect("a mismatched reply channel must be refused");
-
-        assert!(error.contains("Snapshot"), "{error}");
-        assert!(error.contains("Unit"), "{error}");
-    }
-
-    #[tokio::test]
-    async fn a_rebuilt_request_answers_on_the_channel_its_command_asked_for() {
-        let command = GuiCommand::DiffWorkspace;
-        let (reply, receiver) = GuiReplySender::channel(command.reply_kind());
-        let request = GuiRequest::from_parts(command, reply).unwrap();
-
-        let GuiRequest::DiffWorkspace { reply } = request else {
-            panic!("a DiffWorkspace command must rebuild a DiffWorkspace request");
-        };
-        reply
-            .send(Ok(std::path::PathBuf::from("workspace/project")))
-            .unwrap();
-
-        assert_eq!(
-            receiver.recv().await.unwrap(),
-            Some(GuiReply::Path(Ok(std::path::PathBuf::from(
-                "workspace/project"
-            ))))
-        );
-    }
-
-    #[tokio::test]
-    async fn a_fire_and_forget_command_has_nothing_to_wait_for() {
-        let command = GuiCommand::Shutdown;
-        let (reply, receiver) = GuiReplySender::channel(command.reply_kind());
-
-        let request = GuiRequest::from_parts(command, reply).unwrap();
-
-        assert!(matches!(request, GuiRequest::Shutdown));
-        assert_eq!(receiver.recv().await.unwrap(), None);
     }
 
     #[test]
